beta 1

+++ src/ua/in/dris4ecoder/Model/ListCollection.java
@@ -15,66 +15,66 @@ public class ListCollection extends AllCollections {
     }
 
     @Override
-    public long add(int measurementCount) {
+    public double add(int measurementCount) {
         final long start = System.currentTimeMillis();
 
         for (int i = 0; i < measurementCount; i++) {
             ((List<Integer>) collection).add(randomGenerator(0, collection.size() - 1), 0);
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long get(int measurementCount) {
+    public double get(int measurementCount) {
         final long start = System.currentTimeMillis();
 
         for (int i = 0; i < measurementCount; i++) {
             ((List<Integer>) collection).get(randomGenerator(0, collection.size() - 1));
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long remove(int measurementCount) {
+    public double remove(int measurementCount) {
         final long start = System.currentTimeMillis();
 
         for (int i = 0; i < measurementCount; i++) {
             collection.remove(randomGenerator(0, collection.size() - 1));
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long contain(int measurementCount) {
+    public double contain(int measurementCount) {
         final long start = System.currentTimeMillis();
 
         for (int i = 0; i < measurementCount; i++) {
             collection.contains(randomGenerator(0, Integer.MAX_VALUE));
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long populate(int startValue, int endValue, int measurementCount) {
+    public double populate(int startValue, int endValue, int measurementCount) {
 
         final long start = System.currentTimeMillis();
 
         for (int j = 0; j < measurementCount; j++) {
             this.collection.clear();
-            while (collection.size() < collectionSize){
+            while (collection.size() < collectionSize) {
                 collection.add(randomGenerator(startValue, endValue));
             }
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long listIteratorAdd(int measurementCount) {
+    public double listIteratorAdd(int measurementCount) {
         ListIterator<Integer> iterator = ((List<Integer>) collection).listIterator();
 
         final long start = System.currentTimeMillis();
@@ -83,11 +83,11 @@ public class ListCollection extends AllCollections {
             iterator.add(0);
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 
     @Override
-    public long listIteratorRemove(int measurementCount) {
+    public double listIteratorRemove(int measurementCount) {
         ListIterator<Integer> iterator = ((List<Integer>) collection).listIterator();
 
         final long start = System.currentTimeMillis();
@@ -97,6 +97,6 @@ public class ListCollection extends AllCollections {
             iterator.remove();
         }
 
-        return (System.currentTimeMillis() - start) / measurementCount;
+        return (double) (System.currentTimeMillis() - start) / measurementCount;
     }
 }
